+++ openfga_sdk/models/assertion.py
@@ -57,8 +57,7 @@ class Assertion(object):
         self._expectation = None
         self.discriminator = None
 
-        if tuple_key is not None:
-            self.tuple_key = tuple_key
+        self.tuple_key = tuple_key
         self.expectation = expectation
 
     @property
@@ -79,6 +78,8 @@ class Assertion(object):
         :param tuple_key: The tuple_key of this Assertion.  # noqa: E501
         :type tuple_key: TupleKey
         """
+        if self.local_vars_configuration.client_side_validation and tuple_key is None:  # noqa: E501
+            raise ValueError("Invalid value for `tuple_key`, must not be `None`")  # noqa: E501
 
         self._tuple_key = tuple_key