+++ openfga_sdk/models/create_store_request.py
@@ -54,8 +54,7 @@ class CreateStoreRequest(object):
         self._name = None
         self.discriminator = None
 
-        if name is not None:
-            self.name = name
+        self.name = name
 
     @property
     def name(self):
@@ -75,6 +74,8 @@ class CreateStoreRequest(object):
         :param name: The name of this CreateStoreRequest.  # noqa: E501
         :type name: str
         """
+        if self.local_vars_configuration.client_side_validation and name is None:  # noqa: E501
+            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501
 
         self._name = name